refactor(app): extract review-panel comment routing
Move activeFileTab, recent, and openComment (with its commentInReview helper)
into prompt-input/comment-routing.ts. The factory consumes useSessionLayout/
useFile/useComments/useSync internally, exposing only what PromptInput needs.

This logic was inside the composer historically but routes clicks into the
review side panel — it's not composer-internal.
