+++ packages/app/src/components/prompt-input.tsx
@@ -36,7 +36,6 @@ import { usePermission } from "@/context/permission"
 import { useLanguage } from "@/context/language"
 import { canUseNativeFilePicker, usePlatform } from "@/context/platform"
 import { useSessionLayout } from "@/pages/session/session-layout"
-import { createSessionTabs } from "@/pages/session/helpers"
 import { promptEnabled, promptProbe } from "@/testing/prompt"
 import { createTextFragment, getCursorPosition, setCursorPosition, setRangeEdge } from "./prompt-input/editor-dom"
 import {
@@ -45,6 +44,7 @@ import {
   parseEditorToParts,
 } from "./prompt-input/editor-serialize"
 import { createEditorImperatives } from "./prompt-input/editor-imperatives"
+import { createCommentRouting } from "./prompt-input/comment-routing"
 import { createPromptAttachments } from "./prompt-input/attachments"
 import { pickAttachments } from "./prompt-input/pick-attachments"
 import { ACCEPTED_FILE_TYPES } from "./prompt-input/files"
@@ -158,77 +158,8 @@ export const PromptInput: Component<PromptInputProps> = (props) => {
     inset,
   })
 
-  const activeFileTab = createSessionTabs({
-    tabs,
-    pathFromTab: files.pathFromTab,
-    normalizeTab: (tab) => (tab.startsWith("file://") ? files.tab(tab) : tab),
-  }).activeFileTab
+  const { recent, openComment } = createCommentRouting({ activeSessionID })
 
-  const commentInReview = (path: string) => {
-    const sessionID = activeSessionID()
-    if (!sessionID) return false
-
-    const diffs = sync.data.session_diff[sessionID]
-    if (!diffs) return false
-    return diffs.some((diff) => diff.file === path)
-  }
-
-  const openComment = (item: { path: string; commentID?: string; commentOrigin?: "review" | "file" }) => {
-    if (!item.commentID) return
-
-    const focus = { file: item.path, id: item.commentID }
-    comments.setActive(focus)
-
-    const queueCommentFocus = (attempts = 6) => {
-      const schedule = (left: number) => {
-        requestAnimationFrame(() => {
-          comments.setFocus({ ...focus })
-          if (left <= 0) return
-          requestAnimationFrame(() => {
-            const current = comments.focus()
-            if (!current) return
-            if (current.file !== focus.file || current.id !== focus.id) return
-            schedule(left - 1)
-          })
-        })
-      }
-
-      schedule(attempts)
-    }
-
-    const wantsReview = item.commentOrigin === "review" || (item.commentOrigin !== "file" && commentInReview(item.path))
-    view().sidePanel.openTab("review")
-    if (wantsReview) {
-      view().sidePanel.explorer.setTab("changes")
-      tabs().setActive("review")
-      queueCommentFocus()
-      return
-    }
-
-    view().sidePanel.explorer.setTab("all")
-    const tab = files.tab(item.path)
-    tabs().open(tab)
-    tabs().setActive(tab)
-    Promise.resolve(files.load(item.path)).finally(() => queueCommentFocus())
-  }
-
-  const recent = createMemo(() => {
-    const all = tabs().all()
-    const active = activeFileTab()
-    const order = active ? [active, ...all.filter((x) => x !== active)] : all
-    const seen = new Set<string>()
-    const paths: string[] = []
-
-    for (const tab of order) {
-      const path = files.pathFromTab(tab)
-      if (!path) continue
-      if (seen.has(path)) continue
-      seen.add(path)
-      paths.push(path)
-    }
-
-    return paths
-  })
   const info = createMemo(() => (activeSessionID() ? sync.session.get(activeSessionID()!) : undefined))
   const status = createMemo(
     () =>
